[Complete] Property 관련 기능 모두 완성

+++ src/main/java/com/home/house/controller/PropertyController.java
@@ -18,6 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.home.house.model.Property;
 import com.home.house.service.PropertyService;
 
+import jakarta.servlet.http.HttpSession;
+
 @RestController
 @RequestMapping("/property")
 @CrossOrigin("*")
@@ -26,7 +28,7 @@ public class PropertyController {
 	@Autowired
 	private PropertyService propertyService;
 
-	// 글목록 가져오기, 페이징
+	// 글목록 가져오기
 	@GetMapping("/getListAll")
 	public ResponseEntity<?> getPropertyList() {
 		try {
@@ -36,6 +38,22 @@ public class PropertyController {
 			return new ResponseEntity<String>("아무 글도 없습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
+	
+	// 내가 쓴 글목록 가져오기
+	@GetMapping("/getListMy")
+	public ResponseEntity<?> getMyPropertyList(HttpSession session) {
+		try {
+			 String userId = (String) session.getAttribute("member");
+	            if (userId == null) {
+	                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인 필요");
+	            }
+			List<Property> properties = propertyService.getMyList(userId);
+			return new ResponseEntity<List<Property>>(properties, HttpStatus.OK);
+		} catch (Exception e) {
+			return new ResponseEntity<String>("아무 글도 없습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+	
 
 	// 글 상세 보기  
 	@GetMapping("/getDetail")
@@ -50,8 +68,14 @@ public class PropertyController {
 
 	// 글 등록
 	@PostMapping
-	public ResponseEntity<String> addProperty(@RequestBody Property property) {
+	public ResponseEntity<String> addProperty(HttpSession session, @RequestBody Property property) {
 		try {
+	        String userId = (String) session.getAttribute("member");
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인 필요");
+            }
+			
+            property.setId(userId);
 			propertyService.add(property);
 			return new ResponseEntity<String>("글 등록 성공", HttpStatus.OK);
 		} catch (Exception e) {
@@ -62,8 +86,13 @@ public class PropertyController {
 
 	// 글 업데이트
 	@PutMapping
-	public ResponseEntity<String> updateProperty(@RequestBody Property property) {
+	public ResponseEntity<String> updateProperty(HttpSession session, @RequestBody Property property) {
 		try {
+			
+	        String userId = (String) session.getAttribute("member");
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인 필요");
+            }
 			propertyService.update(property);
 			return new ResponseEntity<String>("글 업데이트 성공", HttpStatus.OK);
 		} catch (Exception e) {
@@ -73,8 +102,19 @@ public class PropertyController {
 
 	// 글 삭제
 	@DeleteMapping("/remove")
-	public ResponseEntity<String> removeProperty(@RequestParam("idx") String idx) {
+	public ResponseEntity<String> removeProperty(HttpSession session, @RequestParam("idx") String idx) {
 		try {
+			 String userId = (String) session.getAttribute("member");
+	            if (userId == null) {
+	                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인 필요");
+	         }
+			
+	        int propertyExists = propertyService.isPropertyExists(idx);    
+	        if (propertyExists == 0) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("없는 글");
+            }    
+	        
+	        
 			propertyService.remove(idx);
 			return new ResponseEntity<String>("글 제거 성공", HttpStatus.OK);
 		} catch (Exception e) {
